# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de la Protectora",
        "description": "Documentación de la API para gestión de mascotas",
        "contact": {
            "email": "isa@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/solicitudes": {
            "get": {
                "tags": [
                    "Solicitudes de Adopción"
                ],
                "summary": "Listar todas las solicitudes de adopción",
                "operationId": "7e61361aa7a24c0dc1b8b569584e7835",
                "responses": {
                    "200": {
                        "description": "Listado de solicitudes de adopción"
                    }
                }
            },
            "post": {
                "tags": [
                    "Solicitudes de Adopción"
                ],
                "summary": "Crear una nueva solicitud de adopción",
                "operationId": "5225435ceee3f52b474cb709095c170c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_id",
                                    "mascota_id",
                                    "fecha_solicitud",
                                    "estado"
                                ],
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mascota_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_solicitud": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-01"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "aprobada",
                                            "rechazada"
                                        ],
                                        "example": "pendiente"
                                    },
                                    "comentario": {
                                        "type": "string",
                                        "example": "Me encantaría adoptar a esta mascota"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitud de adopción creada"
                    }
                }
            }
        },
        "/api/solicitudes/paginadas": {
            "get": {
                "tags": [
                    "Solicitudes de Adopción"
                ],
                "summary": "Obtener solicitudes de adopción paginadas",
                "operationId": "00d819997b825326de0a5fcf0baf3771",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de solicitudes de adopción"
                    }
                }
            }
        },
        "/api/solicitudes/{id}": {
            "get": {
                "tags": [
                    "Solicitudes de Adopción"
                ],
                "summary": "Obtener una solicitud de adopción específica",
                "operationId": "bf35f3383ce95b47f2011fb2631a7b71",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud de adopción",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la solicitud de adopción"
                    },
                    "404": {
                        "description": "Solicitud no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Solicitudes de Adopción"
                ],
                "summary": "Actualizar una solicitud de adopción",
                "operationId": "11545b5026dc0d26a78d523ba1324065",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mascota_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_solicitud": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-01"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "aprobada",
                                            "rechazada"
                                        ],
                                        "example": "aprobada"
                                    },
                                    "comentario": {
                                        "type": "string",
                                        "example": "Comentario actualizado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Solicitud de adopción actualizada"
                    },
                    "404": {
                        "description": "Solicitud no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Solicitudes de Adopción"
                ],
                "summary": "Eliminar una solicitud de adopción",
                "operationId": "d456a4f35791639e26c4067da51a68c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud de adopción eliminada"
                    },
                    "404": {
                        "description": "Solicitud no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}